Add a Frequent Network finder.
This automatically extracts the frequent network from the GTFS data,
for a definition of frequent derived from the one used in Vancouver, BC.
This means service running every 15 minutes or better, between 6 am and
9 pm (7 am on Saturdays, 8 am on Sundays). This is not a hard criterion:
there is an allowance for a limited amount of deviation from the norm,
with a sum-of-squares metric for headways that exceed this norm.
These are all computed on a per-stop basis, and a route is considered
frequent if it has more than one stop with frequent service.