Split trip classification into its own file
As it turns out, most feeds have a reliable direction_id field that makes
the somewhat unreliable classification unnecessary. Therefore, it goes into
an as-yet unused classify.py module that will be used to try and reconstruct
direction_id for feeds that don't have it as part of the parsing step.