Add a utility to clean calendar and calendar_dates
Because apparently some transit agencies love to publish a calendar
where there's a Monday to Friday service that happens to not run on
any Friday. Or one that doesn't run on any Monday, Tuesday, Wednesday,
or Thursday. This detects and fixes those, making for a much more
compact representation of service calendar and exceptions.